Move PopupHandler to own file
